backend/requirements.txt
app-settings==1.4.4
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.4
django-allauth==0.58.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
joblib==1.3.2
numpy==1.26.2
joblib==1.3.2
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.3
pandas==2.1.3
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
standard-scaler==0.3
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
app-settings==1.4.4
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.4
django-allauth==0.58.2
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.4
joblib==1.3.2
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.3
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
standard-scaler==0.3
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
